Simit driver moved from development to simit
