--- src/drivers/development/test.seq
@@ -1,4 +1,0 @@
-	IW0	testinput	
-	I2.0	testBoolInput	
-	I2.1	testBoolInput1	
-	QW4	testoutput
