+++ src/drivers/simit/__init__.py
@@ -1,0 +1,1 @@
+from .simit import simit
